:truck: Move & rename localStorage & cookie services
